requirements.txt
supabase==2.10.0
python-dotenv==1.2.2
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.30.0
httpx>=0.27.0
